+++ states/statemgr/lock_test.go
@@ -1,10 +1,10 @@
-package state
+package statemgr
 
 import (
 	"testing"
 )
 
 func TestLockDisabled_impl(t *testing.T) {
-	var _ State = new(LockDisabled)
+	var _ Full = new(LockDisabled)
 	var _ Locker = new(LockDisabled)
 }
